QMP-1: ABM eventos, falta hacer el trabajo de revisarlos periodicamente

+++ src/main/java/utn/frba/dds/que_me_pongo/Model/Evento.java
@@ -27,8 +27,8 @@ public class Evento {
     String nombre;
     Date desde;
     Date hasta;
-    float latitud;
-    float longitud;
+    Double latitud;
+    Double longitud;
     String frecuencia;
 
     @OneToOne(optional = true)
